pyproject.toml
[tool.poetry]
name = "upgraded-journey"
version = "0.1.0"
description = ""
authors = ["Nyaaa <1457047+Nyaaa@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.21.1"
fastapi = "^0.95.0"
asyncpg = "^0.27.0"
psycopg2-binary = "^2.9.6"
sqlalchemy = "^2.0.9"
python-multipart = "^0.0.6"
aiofiles = "^23.1.0"
pydantic = {extras = ["email"], version = "^1.10.7"}
phonenumbers = "^8.13.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
fastapi-versionizer = "^0.1.4"


[tool.poetry.group.dev.dependencies]
httpx = "^0.23.3"
pytest-cov = "^4.0.0"
mock = "^5.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
